deploy/requirements.txt
aws-cdk.core
aws-cdk.aws-events-targets
aws-cdk.aws-lambda
aws_cdk.aws-events
aws_cdk.aws-iam